requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile
#
aerich==0.7.2
    # via deadwood-bot (pyproject.toml)
aiofiles==0.8.0
    # via asyncpraw
aiohttp==3.8.6
    # via
    #   asyncpraw
    #   asyncprawcore
    #   deadwood-bot (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.17.0
    # via
    #   asyncpraw
    #   tortoise-orm
annotated-types==0.5.0
    # via pydantic
async-timeout==4.0.3
    # via aiohttp
asyncpraw==7.7.1
    # via deadwood-bot (pyproject.toml)
asyncprawcore==2.3.0
    # via
    #   asyncpraw
    #   deadwood-bot (pyproject.toml)
asynctest==0.13.0
    # via aiohttp
attrs==23.1.0
    # via aiohttp
cashews==5.3.2
    # via deadwood-bot (pyproject.toml)
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.1
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via aerich
dictdiffer==0.9.0
    # via aerich
diskcache==5.6.3
    # via deadwood-bot (pyproject.toml)
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.7.0
    # via
    #   attrs
    #   click
iso8601==1.1.0
    # via tortoise-orm
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pyaes==1.6.1
    # via pyrogram
pydantic==2.4.2
    # via
    #   aerich
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via deadwood-bot (pyproject.toml)
pykeyboard==0.1.5
    # via deadwood-bot (pyproject.toml)
pypika-tortoise==0.1.6
    # via tortoise-orm
pyrogram==2.0.106
    # via
    #   deadwood-bot (pyproject.toml)
    #   pykeyboard
pysocks==1.7.1
    # via pyrogram
python-dotenv==0.21.1
    # via
    #   deadwood-bot (pyproject.toml)
    #   pydantic-settings
pytz==2023.3.post1
    # via tortoise-orm
requests==2.31.0
    # via update-checker
tgcrypto==1.2.5
    # via
    #   deadwood-bot (pyproject.toml)
    #   pykeyboard
tomlkit==0.12.1
    # via aerich
tortoise-orm==0.19.3
    # via aerich
typing-extensions==4.7.1
    # via
    #   aiohttp
    #   aiosqlite
    #   annotated-types
    #   async-timeout
    #   cashews
    #   importlib-metadata
    #   pydantic
    #   pydantic-core
    #   yarl
update-checker==0.18.0
    # via asyncpraw
urllib3==2.0.7
    # via requests
yarl==1.9.2
    # via
    #   aiohttp
    #   asyncprawcore
zipp==3.15.0
    # via importlib-metadata
